requirements.txt
openai
langchain
langchain-openai
langchain-community
python-dotenv
